test.fix: バリエーションテストについてデータテーブルを使うように修正
* ブロックラベルにコメントもしてみた

+++ src/test/groovy/com/naosim/dddwork/kintai/AppTest.groovy
@@ -1,332 +1,67 @@
 package com.naosim.dddwork.kintai
 
 import spock.lang.Specification
+import spock.lang.Unroll
 
 class AppTest extends Specification {
 
     def CSV_FILE_NAME = "data.csv"
 
-    def "勤務時間-早退-午前"() {
+    @Unroll
+    def "勤務時間-バリエーション: #caseName"() {
 
-        setup:
+        setup: "データファイルが存在しない状態で"
         new File(CSV_FILE_NAME).delete()
         def outputStream = new ByteArrayOutputStream()
         def printStream = Spy(PrintStream, constructorArgs: [outputStream])
         System.out = printStream
 
-        when:
-        App.main("input", "20170101", "0900", "1100")
+        when: "日次勤怠を1件登録すると"
+        App.main("input", date, begin, end)
 
-        then:
-        App.main("total", "201701")
-        outputStream.toString() == """|勤務時間: 2時間0分
-                                      |残業時間: 0時間0分
+        then: "月次勤務時間合計としては、登録した1件分の合計が得られる"
+        App.main("total", date.substring(0, 6))
+        outputStream.toString() == """|勤務時間: ${勤務H}時間${勤務M}分
+                                      |残業時間: ${残業H}時間${残業M}分
                                       |""".stripMargin()
-    }
 
-    def "勤務時間-早退-昼休み-境界begin"() {
-
-        setup:
-        new File(CSV_FILE_NAME).delete()
-        def outputStream = new ByteArrayOutputStream()
-        def printStream = Spy(PrintStream, constructorArgs: [outputStream])
-        System.out = printStream
-
-        when:
-        App.main("input", "20170101", "0900", "1200")
-
-        then:
-        App.main("total", "201701")
-        outputStream.toString() == """|勤務時間: 3時間0分
-                                      |残業時間: 0時間0分
-                                      |""".stripMargin()
-    }
-
-    def "勤務時間-早退-昼休み-中間"() {
-
-        setup:
-        new File(CSV_FILE_NAME).delete()
-        def outputStream = new ByteArrayOutputStream()
-        def printStream = Spy(PrintStream, constructorArgs: [outputStream])
-        System.out = printStream
-
-        when:
-        App.main("input", "20170101", "0900", "1220")
-
-        then:
-        App.main("total", "201701")
-        outputStream.toString() == """|勤務時間: 3時間0分
-                                      |残業時間: 0時間0分
-                                      |""".stripMargin()
-    }
-
-    def "勤務時間-早退-昼休み-境界end"() {
-
-        setup:
-        new File(CSV_FILE_NAME).delete()
-        def outputStream = new ByteArrayOutputStream()
-        def printStream = Spy(PrintStream, constructorArgs: [outputStream])
-        System.out = printStream
-
-        when:
-        App.main("input", "20170101", "0900", "1300")
-
-        then:
-        App.main("total", "201701")
-        outputStream.toString() == """|勤務時間: 3時間0分
-                                      |残業時間: 0時間0分
-                                      |""".stripMargin()
-    }
-
-    def "勤務時間-早退-午後"() {
-
-        setup:
-        new File(CSV_FILE_NAME).delete()
-        def outputStream = new ByteArrayOutputStream()
-        def printStream = Spy(PrintStream, constructorArgs: [outputStream])
-        System.out = printStream
-
-        when:
-        App.main("input", "20170101", "0900", "1540")
-
-        then:
-        App.main("total", "201701")
-        outputStream.toString() == """|勤務時間: 5時間40分
-                                      |残業時間: 0時間0分
-                                      |""".stripMargin()
-    }
-
-    def "勤務時間-定時(夕休み-境界begin)"() {
-
-        setup:
-        new File(CSV_FILE_NAME).delete()
-        def outputStream = new ByteArrayOutputStream()
-        def printStream = Spy(PrintStream, constructorArgs: [outputStream])
-        System.out = printStream
-
-        when:
-        App.main("input", "20170101", "0900", "1800")
-
-        then:
-        App.main("total", "201701")
-        outputStream.toString() == """|勤務時間: 8時間0分
-                                      |残業時間: 0時間0分
-                                      |""".stripMargin()
-    }
-
-    def "勤務時間-定時-夕休み-中間"() {
-
-        setup:
-        new File(CSV_FILE_NAME).delete()
-        def outputStream = new ByteArrayOutputStream()
-        def printStream = Spy(PrintStream, constructorArgs: [outputStream])
-        System.out = printStream
-
-        when:
-        App.main("input", "20170101", "0900", "1810")
-
-        then:
-        App.main("total", "201701")
-        outputStream.toString() == """|勤務時間: 8時間0分
-                                      |残業時間: 0時間0分
-                                      |""".stripMargin()
-    }
-
-    def "勤務時間-定時-夕休み-境界end"() {
-
-        setup:
-        new File(CSV_FILE_NAME).delete()
-        def outputStream = new ByteArrayOutputStream()
-        def printStream = Spy(PrintStream, constructorArgs: [outputStream])
-        System.out = printStream
-
-        when:
-        App.main("input", "20170101", "0900", "1900")
-
-        then:
-        App.main("total", "201701")
-        outputStream.toString() == """|勤務時間: 8時間0分
-                                      |残業時間: 0時間0分
-                                      |""".stripMargin()
-    }
-
-    def "勤務時間-残業-晩"() {
-
-        setup:
-        new File(CSV_FILE_NAME).delete()
-        def outputStream = new ByteArrayOutputStream()
-        def printStream = Spy(PrintStream, constructorArgs: [outputStream])
-        System.out = printStream
-
-        when:
-        App.main("input", "20170101", "0900", "2010")
-
-        then:
-        App.main("total", "201701")
-        outputStream.toString() == """|勤務時間: 9時間10分
-                                      |残業時間: 1時間10分
-                                      |""".stripMargin()
-    }
-
-    def "勤務時間-残業-夜休み-境界begin"() {
-
-        setup:
-        new File(CSV_FILE_NAME).delete()
-        def outputStream = new ByteArrayOutputStream()
-        def printStream = Spy(PrintStream, constructorArgs: [outputStream])
-        System.out = printStream
-
-        when:
-        App.main("input", "20170101", "0900", "2100")
-
-        then:
-        App.main("total", "201701")
-        outputStream.toString() == """|勤務時間: 10時間0分
-                                      |残業時間: 2時間0分
-                                      |""".stripMargin()
-    }
-
-    def "勤務時間-残業-夜休み-中間"() {
-
-        setup:
-        new File(CSV_FILE_NAME).delete()
-        def outputStream = new ByteArrayOutputStream()
-        def printStream = Spy(PrintStream, constructorArgs: [outputStream])
-        System.out = printStream
-
-        when:
-        App.main("input", "20170101", "0900", "2150")
-
-        then:
-        App.main("total", "201701")
-        outputStream.toString() == """|勤務時間: 10時間0分
-                                      |残業時間: 2時間0分
-                                      |""".stripMargin()
-    }
-
-    def "勤務時間-残業-夜休み-境界end"() {
-
-        setup:
-        new File(CSV_FILE_NAME).delete()
-        def outputStream = new ByteArrayOutputStream()
-        def printStream = Spy(PrintStream, constructorArgs: [outputStream])
-        System.out = printStream
-
-        when:
-        App.main("input", "20170101", "0900", "2200")
-
-        then:
-        App.main("total", "201701")
-        outputStream.toString() == """|勤務時間: 10時間0分
-                                      |残業時間: 2時間0分
-                                      |""".stripMargin()
-    }
-
-    def "勤務時間-残業-深夜"() {
-
-        setup:
-        new File(CSV_FILE_NAME).delete()
-        def outputStream = new ByteArrayOutputStream()
-        def printStream = Spy(PrintStream, constructorArgs: [outputStream])
-        System.out = printStream
-
-        when:
-        App.main("input", "20170101", "0900", "2330")
-
-        then:
-        App.main("total", "201701")
-        outputStream.toString() == """|勤務時間: 11時間30分
-                                      |残業時間: 3時間30分
-                                      |""".stripMargin()
-    }
-
-    def "勤務時間-サービス残業-境界begin"() {
-
-        setup:
-        new File(CSV_FILE_NAME).delete()
-        def outputStream = new ByteArrayOutputStream()
-        def printStream = Spy(PrintStream, constructorArgs: [outputStream])
-        System.out = printStream
-
-        when:
-        App.main("input", "20170101", "0900", "2400")
-
-        then:
-        App.main("total", "201701")
-        outputStream.toString() == """|勤務時間: 12時間0分
-                                      |残業時間: 4時間0分
-                                      |""".stripMargin()
-    }
-
-    def "勤務時間-サービス残業-中間1"() {
-
-        setup:
-        new File(CSV_FILE_NAME).delete()
-        def outputStream = new ByteArrayOutputStream()
-        def printStream = Spy(PrintStream, constructorArgs: [outputStream])
-        System.out = printStream
-
-        when:
-        App.main("input", "20170101", "0900", "2530")
-
-        then:
-        App.main("total", "201701")
-        outputStream.toString() == """|勤務時間: 12時間0分
-                                      |残業時間: 4時間0分
-                                      |""".stripMargin()
-    }
-
-    def "勤務時間-サービス残業-中間2"() {
-
-        setup:
-        new File(CSV_FILE_NAME).delete()
-        def outputStream = new ByteArrayOutputStream()
-        def printStream = Spy(PrintStream, constructorArgs: [outputStream])
-        System.out = printStream
-
-        when:
-        App.main("input", "20170101", "0900", "3100")
-
-        then:
-        App.main("total", "201701")
-        outputStream.toString() == """|勤務時間: 12時間0分
-                                      |残業時間: 4時間0分
-                                      |""".stripMargin()
-    }
-
-    def "勤務時間-サービス残業-境界end"() {
-
-        setup:
-        new File(CSV_FILE_NAME).delete()
-        def outputStream = new ByteArrayOutputStream()
-        def printStream = Spy(PrintStream, constructorArgs: [outputStream])
-        System.out = printStream
-
-        when:
-        App.main("input", "20170101", "0900", "3300")
-
-        then:
-        App.main("total", "201701")
-        outputStream.toString() == """|勤務時間: 12時間0分
-                                      |残業時間: 4時間0分
-                                      |""".stripMargin()
+        where:
+         date      | begin | end   || 勤務H | 勤務M | 残業H | 残業M || caseName
+        "20170101" |"0900" |"1100" || 2     | 0    | 0    | 0     || "勤務時間-早退-午前"
+        "20170101" |"0900" |"1100" || 2     | 0    | 0    | 0     || "勤務時間-早退-午前"
+        "20170101" |"0900" |"1200" || 3     | 0    | 0    | 0     || "勤務時間-早退-昼休み-境界begin"
+        "20170101" |"0900" |"1220" || 3     | 0    | 0    | 0     || "勤務時間-早退-昼休み-中間"
+        "20170101" |"0900" |"1300" || 3     | 0    | 0    | 0     || "勤務時間-早退-昼休み-境界end"
+        "20170101" |"0900" |"1540" || 5     | 40   | 0    | 0     || "勤務時間-早退-午後"
+        "20170101" |"0900" |"1800" || 8     | 0    | 0    | 0     || "勤務時間-定時(夕休み-境界begin)"
+        "20170101" |"0900" |"1810" || 8     | 0    | 0    | 0     || "勤務時間-定時-夕休み-中間"
+        "20170101" |"0900" |"1900" || 8     | 0    | 0    | 0     || "勤務時間-定時-夕休み-境界end"
+        "20170101" |"0900" |"2010" || 9     | 10   | 1    | 10    || "勤務時間-残業-晩"
+        "20170101" |"0900" |"2100" || 10    | 0    | 2    | 0     || "勤務時間-残業-夜休み-境界begin"
+        "20170101" |"0900" |"2150" || 10    | 0    | 2    | 0     || "勤務時間-残業-夜休み-中間"
+        "20170101" |"0900" |"2200" || 10    | 0    | 2    | 0     || "勤務時間-残業-夜休み-境界end"
+        "20170101" |"0900" |"2330" || 11    | 30   | 3    | 30    || "勤務時間-残業-深夜"
+        "20170101" |"0900" |"2400" || 12    | 0    | 4    | 0     || "勤務時間-サービス残業-境界begin"
+        "20170101" |"0900" |"2530" || 12    | 0    | 4    | 0     || "勤務時間-サービス残業-中間1"
+        "20170101" |"0900" |"3100" || 12    | 0    | 4    | 0     || "勤務時間-サービス残業-中間2"
+        "20170101" |"0900" |"2530" || 12    | 0    | 4    | 0     || "勤務時間-サービス残業-境界end"
     }
 
 
 
     def "修正登録-1回修正"() {
 
-        setup:
+        setup: "データファイルが存在しない状態で"
         new File(CSV_FILE_NAME).delete()
         def outputStream = new ByteArrayOutputStream()
         def printStream = Spy(PrintStream, constructorArgs: [outputStream])
         System.out = printStream
 
-        when:
+        when: "同じ勤怠日付で2回登録すると"
         App.main("input", "20170101", "0900", "1800")
         App.main("input", "20170101", "0900", "2030")
 
-        then:
+        then: "最後に登録したデータが有効になる"
         App.main("total", "201701")
         outputStream.toString() == """|勤務時間: 9時間30分
                                       |残業時間: 1時間30分
@@ -335,18 +70,18 @@ class AppTest extends Specification {
 
     def "修正登録-2回修正"() {
 
-        setup:
+        setup: "データファイルが存在しない状態で"
         new File(CSV_FILE_NAME).delete()
         def outputStream = new ByteArrayOutputStream()
         def printStream = Spy(PrintStream, constructorArgs: [outputStream])
         System.out = printStream
 
-        when:
+        when: "同じ勤怠日付で3回登録すると"
         App.main("input", "20170101", "0900", "1800")
         App.main("input", "20170101", "0900", "2030")
         App.main("input", "20170101", "0900", "1540")
 
-        then:
+        then: "最後に登録したデータが有効になる"
         App.main("total", "201701")
         outputStream.toString() == """|勤務時間: 5時間40分
                                       |残業時間: 0時間0分
@@ -357,13 +92,13 @@ class AppTest extends Specification {
 
     def "シナリオ-月合計-月抽出-修正入力あり"() {
 
-        setup:
+        setup: "データファイルが存在しない状態で"
         new File(CSV_FILE_NAME).delete()
         def outputStream = new ByteArrayOutputStream()
         def printStream = Spy(PrintStream, constructorArgs: [outputStream])
         System.out = printStream
 
-        when:
+        when: "複数の月が存在するように登録すると"
         App.main("input", "20161231", "0900", "1800")   // total対象外
         App.main("input", "20170101", "0900", "1100")
         App.main("input", "20170102", "0900", "1130")
@@ -386,7 +121,7 @@ class AppTest extends Specification {
         App.main("input", "20170131", "0900", "3300")
         App.main("input", "20170201", "0900", "1800")   // total対象外
 
-        then:
+        then: "指定された月についての合計が得られる"
         App.main("total", "201701")
         outputStream.toString() == """|勤務時間: 139時間20分
                                       |残業時間: 26時間40分
